+++ e2e-tests/page-objects/base-window.ts
@@ -1,5 +1,9 @@
 export abstract class BaseWindow {
   protected static async baseActivate(windowUrl: string): Promise<void> {
+    if (await browser.getUrl() === windowUrl) {
+      return;
+    }
+
     const maxRetries = 5;
     const retryDelayMs = 500;
     const urls: string[] = [];