+++ e2e-tests/page-objects/main-window/plugins-view/plugins-view.ts
@@ -1,0 +1,7 @@
+export abstract class PluginsView {
+  private static readonly selfSelector = $('<jcm-plugins-navigation-view />').nextElement();
+
+  static get installedTabButton(): ChainablePromiseElement {
+    return this.selfSelector.$('[data-testid="installed"]');
+  }
+}